fix: avoid uint32 truncation for large validation weights
- Change ShouldValidate parameters from uint32 to int64 to prevent truncation
  at scale when weights exceed 4.29 billion
- Add division-by-zero guard when totalPower == executorPower
- Update all callers to use int64 directly instead of casting to uint32
- Add tests for division-by-zero and large weight scenarios

Fixes #646
Related to #625

+++ decentralized-api/internal/validation/inference_validation.go
@@ -105,9 +105,9 @@ func (s *InferenceValidator) shouldValidateInference(
 	shouldValidate, message := calculations.ShouldValidate(
 		seed,
 		inferenceDetails,
-		uint32(inferenceDetails.TotalPower),
-		uint32(validatorPower),
-		uint32(inferenceDetails.ExecutorPower),
+		int64(inferenceDetails.TotalPower),
+		int64(validatorPower),
+		int64(inferenceDetails.ExecutorPower),
 		validationParams,
 		false)
